Clean up bindings code a touch
