+++ .github/workflows/integration.yml
@@ -29,6 +29,14 @@ jobs:
         os: ["ubuntu-latest", "windows-latest", "macos-latest"]
         rust: ["stable"]
         ruby: ["2.7", "3.2", "ruby-head"]
+        # Should re-include these after stable-abi is enabled on magnus
+        exclude:
+          - os: windows-latest
+            ruby: ruby-head
+            slug: magnus-head
+          - os: windows-latest
+            ruby: ruby-head
+            slug: magnus-0.5
 
     runs-on: ${{ matrix.os }}
     steps:
