+++ crates/rb-sys-build/src/bindings.rs
@@ -1,6 +1,6 @@
 mod sanitizer;
+mod stable_abi;
 
-use crate::bindings::sanitizer::categorize_by_abi_stability;
 use crate::utils::is_msvc;
 use crate::RbConfig;
 use quote::ToTokens;
@@ -68,8 +68,7 @@ pub fn generate(
             .blocklist_item("^_bindgen_ty_9.*")
     };
 
-    let bindings = gen_opaque_struct(bindings, "RArray", &mut wrapper_h);
-    let bindings = gen_opaque_struct(bindings, "RString", &mut wrapper_h);
+    let bindings = stable_abi::opaqueify_bindings(bindings, &mut wrapper_h);
 
     let mut tokens = {
         let bindings = bindings.header_contents("wrapper.h", &wrapper_h);
@@ -95,7 +94,7 @@ pub fn generate(
         }
 
         push_cargo_cfg_from_bindings(&tokens, cfg_out).expect("write cfg");
-        categorize_by_abi_stability(&mut tokens);
+        stable_abi::categorize_bindings(&mut tokens);
         tokens.into_token_stream().to_string()
     };
 
@@ -249,20 +248,4 @@ impl<'a> ConfValue<'a> {
             ),
         }
     }
-}
-
-fn gen_opaque_struct(
-    bindings: bindgen::Builder,
-    name: &str,
-    wrapper_h: &mut String,
-) -> bindgen::Builder {
-    let struct_name = format!("rb_sys__Opaque__{}", name);
-    wrapper_h.push_str(&format!(
-        "struct {} {{ struct {} dummy; }};\n",
-        struct_name, name
-    ));
-
-    bindings
-        .opaque_type(&struct_name)
-        .allowlist_type(struct_name)
 }
